requirements.txt
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
sqlite-vec>=0.1.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
openai>=1.0.0
anthropic>=0.7.0
requests>=2.31.0
psutil>=5.9.0
